+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix 5.9 related WSOD
 * Disable air-helper media directory hook as it breaks things in production
 * Fix hero block labels in front end
+* newblock.sh block importer: Add accordion
 
 ### 1.0.3: 2021-02-08